nyay_backend/requirements.txt
# Django backend stack
Django>=5.0
djangorestframework>=3.14.0

# NyaySetu intelligence dependencies
chromadb>=0.4.22
transformers>=4.36.0
torch>=2.1.0
sentence-transformers>=2.2.2
PyPDF2>=3.0.0
pdfplumber>=0.10.0
openai>=1.6.0
numpy>=1.24.0
regex>=2023.12.25
